+++ Include/Misra/Parsers.h
@@ -1,0 +1,17 @@
+// Filename: misra/parsers
+// Author: Siddharth Mishra <admin@brightprogrammer.in>
+//
+// Copyright (c) 2025 Siddharth Mishra . All rights reserved.
+//
+// This source code is the intellectual property of the author.
+// Redistribution or use, in whole or in part, with or without
+// modification, is strictly prohibited without prior written permission.
+
+
+#ifndef MISRA_PARSERS_H
+#define MISRA_PARSERS_H
+
+#include <Misra/Parsers/C.h>
+#include <Misra/Parsers/JSON.h>
+
+#endif // MISRA_PARSERS_H
