+++ meson.build
@@ -14,7 +14,8 @@ misra_std_sources = files(
     'Source/Misra/Std/File.c',
     'Source/Misra/Std/Container/Vec.c',
     'Source/Misra/Std/Container/Str.c',
-    'Source/Misra/Parsers/JSON.c'
+    'Source/Misra/Parsers/JSON.c',
+    'Source/Misra/Parsers/C.c'
 )
 
 misra_std = static_library(
